Adding tons of Cell tests with metrics
Until I figure out how to match the C++ Subdivide, I've opted to use
CellFromCellID on cell children. Supposedly slower but at least it works.

Out of all this work I only have a _single_ test case failing and I'm
not sure why. It's the MaxArea metric and the difference is tiny
(e.g. 9.36424007478522e-15 > 9.364012772742995e-15).

I also updated the metrics to use the quadratic projections since that's
what stToUV() uses.

+++ s2/cellunion.go
@@ -145,5 +145,5 @@ func (cu CellUnion) IntersectsCellID(id CellID) bool {
 type byID []CellID
 
 func (cu byID) Len() int           { return len(cu) }
-func (cu byID) Less(i, j int) bool { return cu[i] < cu[j] }
+func (cu byID) Less(i, j int) bool { return uint64(cu[i]) < uint64(cu[j]) }
 func (cu byID) Swap(i, j int)      { cu[i], cu[j] = cu[j], cu[i] }
